Added Solution
In the android.rb it was created two new functions 'resetar' and 'biblioteca_de_registros'. Furthermore, it was updated the 'numero' to make the serial number correctly.

In android_test.rb, only one modification, it was added a variable in 'resetar' to check if exist or not that serial number